refactor: 디렉토리 변경사항 수정
