--- src/components/shared/Logout.js
@@ -1,8 +1,0 @@
-import { useNavigate } from "react-router-dom";
-
-const Logout = () => {
-  localStorage.clear();
-  window.location.replace("/");
-};
-
-export default Logout;
